+++ src/notification-manager.ts
@@ -1,39 +1,84 @@
-class NotificationManager {
-    constructor() {
-        this.handleCloseClickEvent = this.removeNotification.bind(this);
-        this.handleActionButtonClick = this.activateButton.bind(this);
+type NotificationButton = {
+    label: string,
+    callback: Function,
+    ariaLabel?: string,
+}
+
+type SnackbarNotification = {
+    message: string,
+    duration?: number,
+    closeable?: boolean,
+    buttons?: Array<NotificationButton>,
+    position?: string,
+    element?: HTMLElement,
+    force?: boolean,
+}
+
+type VerificationResponse = {
+    validNotification: SnackbarNotification,
+    warnings: Array<string>,
+}
+
+export class NotificationManager
+{
+    private _queue: Array<SnackbarNotification>;
+    private _callback: Function;
+    private _isRunning: boolean;
+    private _time: number;
+
+    constructor()
+    {
         this._queue = [];
-        this._callback = () => { };
+        this._callback = ()=>{};
         this._isRunning = false;
         this._time = 0;
     }
-    activateButton(e) {
-        const buttonEl = e.currentTarget;
+
+    private handleCloseClickEvent:EventListener = this.removeNotification.bind(this);
+    private handleActionButtonClick:EventListener = this.activateButton.bind(this);
+
+    private activateButton(e:Event) : void
+    {
+        const buttonEl = e.currentTarget as HTMLButtonElement;
         const button = this._queue[0].buttons[parseInt(buttonEl.dataset.index)];
         button.callback();
         this.removeNotification();
     }
-    createNotification(notification) {
+
+    private createNotification(notification:SnackbarNotification) : void
+    {
         const el = document.createElement('snackbar-component');
         el.setAttribute('position', notification.position);
+        
         const message = document.createElement('p');
         message.innerText = notification.message;
         el.appendChild(message);
-        if (notification.closeable || notification.buttons.length) {
+
+        if (notification.closeable || notification.buttons.length)
+        {
             const actionsWrapper = document.createElement('snackbar-actions');
-            if (notification.buttons.length) {
-                for (let i = 0; i < notification.buttons.length; i++) {
+
+            if (notification.buttons.length)
+            {
+                for (let i = 0; i < notification.buttons.length; i++)
+                {
                     const button = document.createElement('button');
                     button.innerText = notification.buttons[i].label;
-                    button.dataset.index = `${i}`;
-                    if (notification.buttons[i].ariaLabel) {
+                    button.dataset.index = `${ i }`;
+                    
+                    if (notification.buttons[i].ariaLabel)
+                    {
                         button.setAttribute('aria-label', notification.buttons[i].ariaLabel);
                     }
+
                     button.addEventListener('click', this.handleActionButtonClick);
+
                     actionsWrapper.appendChild(button);
                 }
             }
-            if (notification.closeable) {
+
+            if (notification.closeable)
+            {
                 const closeButton = document.createElement('close-button');
                 closeButton.setAttribute('aria-label', 'close notification');
                 closeButton.setAttribute('aria-pressed', 'false');
@@ -42,184 +87,262 @@ class NotificationManager {
                 closeButton.addEventListener('click', this.handleCloseClickEvent);
                 actionsWrapper.appendChild(closeButton);
             }
+
             el.appendChild(actionsWrapper);
         }
+
         document.body.appendChild(el);
         notification.element = el;
     }
-    removeNotification() {
+
+    private removeNotification() : void
+    {
         const closeButton = this._queue[0].element.querySelector('close-button');
-        if (closeButton) {
+        if (closeButton)
+        {
             closeButton.removeEventListener('click', this.handleCloseClickEvent);
         }
+
         const buttons = Array.from(this._queue[0].element.querySelectorAll('button'));
-        if (buttons.length) {
-            for (let i = 0; i < buttons.length; i++) {
+        if (buttons.length)
+        {
+            for (let i = 0; i < buttons.length; i++)
+            {
                 buttons[i].removeEventListener('click', this.activateButton);
             }
         }
+
         this._queue[0].element.remove();
-        this._queue.splice(0, 1);
-        if (this._queue.length !== 0) {
+        this._queue.splice(0,1);
+        if (this._queue.length !== 0)
+        {
             this.createNotification(this._queue[0]);
         }
-        else {
+        else
+        {
             this.stopCallback();
         }
     }
-    startCallback() {
-        if (this._isRunning) {
+
+    private startCallback() : void
+    {
+        if (this._isRunning)
+        {
             return;
         }
+
         this._isRunning = true;
         this._callback = this.animationFrameCallback.bind(this);
         this._time = performance.now();
         this._callback();
         this.createNotification(this._queue[0]);
     }
-    stopCallback() {
+
+    private stopCallback() : void
+    {
         this._isRunning = false;
-        this._callback = () => { };
+        this._callback = ()=>{};
     }
-    animationFrameCallback() {
-        if (this._queue.length === 0 || !this._isRunning) {
+
+    private animationFrameCallback() : void
+    {
+        if (this._queue.length === 0 || !this._isRunning)
+        {
             this.stopCallback();
             return;
         }
+
         const newTime = performance.now();
         const deltaTime = (newTime - this._time) / 1000;
         this._time = newTime;
-        if (document.hasFocus() && this._queue[0].duration !== Infinity) {
+
+        if (document.hasFocus() && this._queue[0].duration !== Infinity)
+        {
             this._queue[0].duration -= deltaTime;
-            if (this._queue[0].duration <= 0) {
+            if (this._queue[0].duration <= 0)
+            {
                 this.removeNotification();
             }
         }
+
         window.requestAnimationFrame(() => { this._callback(); });
     }
-    validateNotification(notification) {
+
+    private validateNotification(notification:SnackbarNotification) : Promise<string|VerificationResponse>
+    {
         return new Promise((resolve, reject) => {
-            const newNotification = {};
-            let warnings = [];
-            if (notification.element) {
+            const newNotification:Partial<SnackbarNotification> = {};
+            let warnings:Array<string> = [];
+
+            if (notification.element)
+            {
                 reject('Notifications create their own HTMLElement, do not provide one.');
             }
-            if (!notification.message) {
+
+            if (!notification.message)
+            {
                 reject('Notifications require a message.');
             }
-            else {
+            else
+            {
                 newNotification.message = notification.message;
             }
-            if (notification.closeable) {
+
+            if (notification.closeable)
+            {
                 newNotification.closeable = notification.closeable;
             }
-            else {
+            else
+            {
                 newNotification.closeable = false;
             }
-            if (notification.duration) {
+
+            if (notification.duration)
+            {
                 newNotification.duration = notification.duration;
-                if (notification.duration === Infinity && newNotification.closeable) {
+
+                if (notification.duration === Infinity && newNotification.closeable)
+                {
                     newNotification.duration = Infinity;
                 }
             }
-            else {
+            else
+            {
                 newNotification.duration = 10;
             }
-            if (notification.position) {
-                if (notification.position.match('top')) {
+
+            if (notification.position)
+            {
+                if (notification.position.match('top'))
+                {
                     newNotification.position = 'top';
                 }
-                else if (notification.position.match('bottom')) {
+                else if (notification.position.match('bottom'))
+                {
                     newNotification.position = 'bottom';
                 }
-                else {
+                else
+                {
                     newNotification.position = 'bottom';
                 }
-                if (notification.position.match('left')) {
+
+                if (notification.position.match('left'))
+                {
                     newNotification.position += ' left';
                 }
-                else if (notification.position.match('right')) {
+                else if (notification.position.match('right'))
+                {
                     newNotification.position += ' right';
                 }
-                else if (notification.position.match('center')) {
+                else if (notification.position.match('center'))
+                {
                     newNotification.position += ' center';
                 }
-                else {
+                else
+                {
                     newNotification.position += ' center';
                 }
             }
-            else {
+            else
+            {
                 newNotification.position = 'bottom center';
             }
-            if (notification.buttons) {
-                const buttons = [];
-                for (let i = 0; i < notification.buttons.length; i++) {
-                    const button = {};
-                    const newWarnings = [];
-                    if (notification.buttons[i].label) {
+
+            if (notification.buttons)
+            {
+                const buttons:Array<NotificationButton> = [];
+
+                for (let i = 0; i < notification.buttons.length; i++)
+                {
+                    const button:Partial<NotificationButton> = {};
+                    const newWarnings:Array<string> = [];
+
+                    if (notification.buttons[i].label)
+                    {
                         button.label = notification.buttons[i].label;
                     }
-                    else {
+                    else
+                    {
                         newWarnings.push('Buttons require a label.');
                     }
-                    if (notification.buttons[i].callback) {
-                        if (typeof notification.buttons[i].callback === 'function') {
+
+                    if (notification.buttons[i].callback)
+                    {
+                        if (typeof notification.buttons[i].callback === 'function')
+                        {
                             button.callback = notification.buttons[i].callback;
                         }
-                        else {
+                        else
+                        {
                             newWarnings.push('Buttons callbacks must be a function.');
                         }
                     }
-                    else {
+                    else
+                    {
                         newWarnings.push('Buttons require a callback function.');
                     }
-                    if (notification.buttons[i].ariaLabel) {
+
+                    if (notification.buttons[i].ariaLabel)
+                    {
                         button.ariaLabel = notification.buttons[i].ariaLabel;
                     }
+
                     warnings = [...warnings, ...newWarnings];
-                    if (newWarnings.length === 0) {
+                    if (newWarnings.length === 0)
+                    {
+                        // @ts-ignore
                         buttons.push(button);
                     }
                 }
-                if (buttons.length) {
+
+                if (buttons.length)
+                {
                     newNotification.buttons = buttons;
                 }
             }
-            else {
+            else
+            {
                 newNotification.buttons = [];
             }
-            if (notification.force) {
+
+            if (notification.force)
+            {
                 newNotification.force = notification.force;
             }
-            else {
+            else
+            {
                 newNotification.force = false;
             }
+
+            // @ts-ignore
             resolve({ validNotification: newNotification, warnings: warnings });
         });
     }
-    notify(notification) {
+
+    public notify(notification:SnackbarNotification) : void
+    {
         this.validateNotification(notification)
-            .then((response) => {
-            if (this._queue.length === 0 || !response.validNotification.force) {
+        .then((response:VerificationResponse) => {
+            if (this._queue.length === 0 || !response.validNotification.force)
+            {
                 this._queue.push(response.validNotification);
             }
-            else if (this._queue.length > 0 && response.validNotification.force) {
+            else if (this._queue.length > 0 && response.validNotification.force)
+            {
                 this._queue.splice(1, 0, response.validNotification);
                 this.removeNotification();
             }
             this.startCallback();
-            if (response.warnings.length !== 0) {
-                for (let i = 0; i < response.warnings.length; i++) {
+            if (response.warnings.length !== 0)
+            {
+                for (let i = 0; i < response.warnings.length; i++)
+                {
                     console.warn(response.warnings[i]);
                 }
             }
         })
-            .catch(error => {
+        .catch(error => {
             console.error(error);
         });
     }
 }
-const globalManager = new NotificationManager();
-const notify = globalManager.notify.bind(globalManager);
-export { NotificationManager, notify };
-//# sourceMappingURL=Notify.js.map