+++ src/notification-manager.ts
@@ -58,18 +58,21 @@ export class NotificationManager {
     private createNotification(notification: SnackbarNotification): void {
         const el = document.createElement("snackbar-component");
 
-        for (let i = 0; i < notification.classes.length; i++) {
+        /** @deprecated */
+        el.setAttribute("position", notification.position);
+
+        for (let i = 0; i < notification?.classes?.length; i++) {
             el.classList.add(notification.classes[i]);
         }
 
         const message = document.createElement("p");
-        message.innerText = notification.message;
+        message.innerText = notification?.message;
         el.appendChild(message);
 
-        if (notification.closeable || notification.buttons.length) {
+        if (notification?.closeable || notification?.buttons?.length) {
             const actionsWrapper = document.createElement("snackbar-actions");
 
-            if (notification.buttons.length) {
+            if (notification?.buttons?.length) {
                 for (let i = 0; i < notification.buttons.length; i++) {
                     const button = document.createElement("button");
                     button.innerText = notification.buttons[i].label;
@@ -89,7 +92,7 @@ export class NotificationManager {
                 }
             }
 
-            if (notification.closeable) {
+            if (notification?.closeable) {
                 const closeButton = document.createElement("close-button");
                 closeButton.setAttribute("aria-label", "close notification");
                 closeButton.setAttribute("aria-pressed", "false");
@@ -108,18 +111,6 @@ export class NotificationManager {
     }
 
     private removeNotification(): void {
-        const closeButton = this._queue[0].element.querySelector("close-button");
-        if (closeButton) {
-            closeButton.removeEventListener("click", this.handleCloseClickEvent);
-        }
-
-        const buttons = Array.from(this._queue[0].element.querySelectorAll("button"));
-        if (buttons.length) {
-            for (let i = 0; i < buttons.length; i++) {
-                buttons[i].removeEventListener("click", this.activateButton);
-            }
-        }
-
         this._queue[0].element.remove();
         this._queue.splice(0, 1);
         if (this._queue.length !== 0) {
@@ -131,7 +122,9 @@ export class NotificationManager {
         this._callback = this.animationFrameCallback.bind(this);
         this._time = performance.now();
         this._callback();
-        this.createNotification(this._queue[0]);
+        if (this._queue.length) {
+            this.createNotification(this._queue[0]);
+        }
     }
 
     private animationFrameCallback(): void {
@@ -140,10 +133,12 @@ export class NotificationManager {
         this._time = newTime;
 
         if (document.hasFocus()) {
-            if (this._queue?.[0]?.duration !== Infinity) {
-                this._queue[0].duration -= deltaTime;
-                if (this._queue[0].duration <= 0) {
-                    this.removeNotification();
+            if (this._queue.length) {
+                if (this._queue[0].duration !== Infinity) {
+                    this._queue[0].duration -= deltaTime;
+                    if (this._queue[0].duration <= 0) {
+                        this.removeNotification();
+                    }
                 }
             }
             if (this.toaster.length) {
@@ -190,6 +185,29 @@ export class NotificationManager {
                 }
             } else {
                 newNotification.duration = 10;
+            }
+
+            if (notification.position) {
+                warnings.push("The position property is deprecated");
+                if (notification.position.match("top")) {
+                    newNotification.position = "top";
+                } else if (notification.position.match("bottom")) {
+                    newNotification.position = "bottom";
+                } else {
+                    newNotification.position = "bottom";
+                }
+
+                if (notification.position.match("left")) {
+                    newNotification.position += " left";
+                } else if (notification.position.match("right")) {
+                    newNotification.position += " right";
+                } else if (notification.position.match("center")) {
+                    newNotification.position += " center";
+                } else {
+                    newNotification.position += " center";
+                }
+            } else {
+                newNotification.position = "bottom center";
             }
 
             if (notification.position) {
@@ -275,7 +293,7 @@ export class NotificationManager {
                 newNotification.force = false;
             }
 
-            if (notification.classes) {
+            if (notification?.classes) {
                 if (typeof notification.classes === "string") {
                     newNotification.classes = [notification.classes];
                 } else if (Array.isArray(notification.classes)) {
@@ -297,6 +315,9 @@ export class NotificationManager {
             .then((response: VerificationResponse) => {
                 if (this._queue.length === 0 || !response.validNotification.force) {
                     this._queue.push(response.validNotification);
+                    if (this._queue.length === 1) {
+                        this.createNotification(this._queue[0]);
+                    }
                 } else if (this._queue.length > 0 && response.validNotification.force) {
                     this._queue.splice(1, 0, response.validNotification);
                     this.removeNotification();
@@ -318,14 +339,14 @@ export class NotificationManager {
     }
     private handleToastClose: EventListener = (e: Event) => {
         const target = e.currentTarget as HTMLButtonElement;
-        const id = parseInt(target.dataset.id);
+        const id = parseInt(target.parentElement.dataset.id);
         this.removeToasterNotification(this.toaster[id], id);
     };
 
     private createToast(notification: ToasterNotification) {
         const el = document.createElement("toast-component");
 
-        for (let i = 0; i < notification.classes.length; i++) {
+        for (let i = 0; i < notification?.classes?.length; i++) {
             el.classList.add(notification.classes[i]);
         }
 
@@ -337,10 +358,10 @@ export class NotificationManager {
 
         const copyWrapper = document.createElement("copy-wrapper");
         const title = document.createElement("h3");
-        title.innerText = notification.title;
+        title.innerText = notification?.title;
         copyWrapper.appendChild(title);
         const message = document.createElement("p");
-        message.innerText = notification.message;
+        message.innerText = notification?.message;
         copyWrapper.appendChild(message);
         el.appendChild(copyWrapper);
 
@@ -366,6 +387,13 @@ export class NotificationManager {
         const notificationEl = this.createToast(notification);
         notificationEl.dataset.id = `${this.toaster.length}`;
         notification.element = notificationEl;
+
+        if (notification?.duration && !isNaN(notification.duration)) {
+            notification.duration = notification.duration;
+        } else {
+            notification.duration = 4;
+        }
+
         this.toaster.push(notification);
         shell.appendChild(notificationEl);
     }